fs: nvs: Add checking for free space not needing garbage collection
The goal is to make sure a determined amount of contiguous free space
is present into the NVS for a subsequent write.

The subsequent write needs to write to the NVS as fast as possible,
so triggering the NVS garbage collector must be avoided.

This is the reason why this function is introduced.
If the GC must be triggered to make some room, it will trigger it,
so the subsequent write will not be slowed by the GC process.

Signed-off-by: Adrien Ricciardi <aricciardi@baylibre.com>