drivers: sensor: ntc_thermistor: adds support for VDD based ADC reference
The NTC thermistor implementation assumes a constant pull-up voltage
and that the ADC channel is measured against a reference voltage so that
the absolute voltage across the NTC can be calculated. Based on the
relationship of `pullup-uv` and this absolute NTC voltage, the resistance
of the NTC is calculated.

There are applications where the "pullup-uv" is not constant, but VDD.
Most ADCs support relative measurements against VDD. If `pullup-uv` is not
defined, the implementation assumes now that the ADC channel is configured
to use VDD as a reference and therefore no millivolt conversion is
required.

In addition, only one of the specified pull-up/pull-down resistances was
taken into account. If a FET switches the resistor divider on the NTC side,
the NTC resistance was too high. In the new implementation, the specified
value (e.g. FET resistance) is subtracted from the NTC to achieve slightly
better accuracy.

Signed-off-by: Stefan Schwendeler <Stefan.Schwendeler@husqvarnagroup.com>