refactor(st_b_lcd40_dsi1_mb1166_a09): disable default STM32_LTDC_RGB888 configuration
I removed the configuration for `STM32_LTDC_RGB888` in the stm32h747i_disco_stm32h747xx_m7.conf, as I believe this should not be set here. When enabling DMA2D (Chrom-ART) this line leads to significant issues (noise on display). Within this situation it needs to be set to `STM32_LTDC_ARGB8888` instead. Within the context of Zephyr, LVGL and the STM32H747I-DISCO many people online mention to be running into problems when activating DMA2D. Deactivating this line solved it for me. Although I'm not an expert I think best practise should be to declare this within the project `prj.conf`.

Signed-off-by: Maarten Scherpenisse <git@maartenscherpenisse.nl>