posix: threads: implement pthread_kill()
Now that POSIX_REALTIME_SIGNALS is implemented, add the final
piece of the POSIX_THREADS puzzle. Namely, pthread_kill().

This was the last function required to mark the
POSIX_THREADS_BASE Option Group as fully supported in Zephyr.

Additionally, we may now mark the _POSIX_THREADS Option as
supported in Zephyr. The _POSIX_THREADS Option is required
as part of the POSIX System Interfaces (Base Definitions)
as well as PSE51, PSE52, PSE53, PSE54 and all other conformant
POSIX subprofiles.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>